+++ tests/testthat/test-clustering.R
@@ -1,10 +1,7 @@
-library(testthat)        # load testthat package
-library(evprof)
-library(dplyr)
-library(tibble)
-library(lubridate)
-library(purrr)
-library(ggplot2)
+
+options(
+  evprof.start.hour = 3
+)
 
 # Get the example `evmodel` and `sessions` included in the package
 ev_model <- evprof::california_ev_model
@@ -59,7 +56,7 @@ test_that("Clusers are plotted correctly", {3
     cluster_sessions(k = 2, seed = 123, log = TRUE)
   plot_clusters <- plot_bivarGMM(sessions_clusters$sessions, sessions_clusters$models, log = FALSE)
   plot_clusters_log <- plot_bivarGMM(sessions_clusters$sessions, sessions_clusters$models, log = TRUE)
-  expect_true(is.ggplot(plot_clusters))
-  expect_true(is.ggplot(plot_clusters_log))
+  expect_true(ggplot2::is.ggplot(plot_clusters))
+  expect_true(ggplot2::is.ggplot(plot_clusters_log))
 })