+++ NEWS.md
@@ -1,6 +1,7 @@
 # evprof 1.1.1
 
 * Improved the consistency of the provided example data sets
+* Function `plot_points` can configure `start` hour when `log = TRUE`
 
 
 # evprof 1.1.0